A2dpSrc: Don't update connection state with dummy device address
While streaming is actively going on, and then BT turn off on DUT,
then HOST is updating the connection state with dummy device BD addr
to stack. But due to dummy address, it is exceeding the connected
indices, which is gicing the 'ubsan: out-of-bounds'.

To avoid this issue, don't update connection state with dummy add.

CRs-Fixed: 2465134
Change-Id: I91e522bb64c88a36fe222cebf0dad1af1428f7ae